% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sangergenotype.R
\name{sangergenotype}
\alias{sangergenotype}
\title{A function to interpret genotype information from ab1 files}
\usage{
sangergenotype(
  dir = choose.dir(),
  link = "",
  wtseq = "",
  mutseq = "",
  revcomp = TRUE,
  cutoff = 0.33,
  pattern = "K[0-9]{4}+"
)
}
\arguments{
\item{dir}{Source directory for .ab1 files to analyze or, if \var{link} is not blank, directory where to download and unpack the archive containing .ab1 files to. In that case, a new subdirectory named by the current date will be created. If left blank, a file open dialog will appear.}

\item{link}{Optional. A url to an archive containing the ab1 files to analyze.}

\item{wtseq}{Sequence snipplet caracteristic for wild-type}

\item{mutseq}{Sequence snipplet caracteristic for mutant}

\item{revcomp}{make reverse complementary of wtseq and mutseq}

\item{cutoff}{minimum relative amplitude of secondary peak. This is the threshold for interpreting a double peak as heterozygous.}

\item{pattern}{A regular expression that matches a unique pattern in the file name (like the mouse name). Could be e.g. \code{"K[0-9]{4}+"} or \code{"[0-9]{1,2}_[0-9]{1,2}[a-z]{1}+"}.}
}
\value{
Data Frame with list of mice and respective genotypes
}
\description{
This function retrieves ab1 files either from a link or from a directory and compares the sequence to a reference sequence. it interprets signal amplitudes in abi traces and can thereby fidderentiate between wt, hom and het.
}
\examples{
\dontrun{
sangergenotype(dir="D:\\\\", link="http://www.x.zip",wtseq="ACTGAAAA",mutseq="ACCGAAAA", revcomp = TRUE, cutoff = 0.2)
} 
}
