% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickAtLab_Import.R
\name{TickAtLab_Import}
\alias{TickAtLab_Import}
\title{TickAtLab_Import}
\usage{
TickAtLab_Import(
  filename = selectFile(caption = "Select tick@lab XLSX File", filter =
    "XLS Files (*.xlsx)", existing = TRUE)
)
}
\arguments{
\item{filename}{Path to tick@lab xls file.}
}
\value{
A data frame containing animal colony information. One row per animal.
}
\description{
Imports xls files created by the Tick@lab export function into a data frame. Only keeps animal names and discardes IDs.
}
