% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickAtLab_pedigree.R
\name{TickAtLab_pedigree}
\alias{TickAtLab_pedigree}
\title{TickAtLab_pedigree}
\usage{
TickAtLab_pedigree(df, features = NULL)
}
\arguments{
\item{df}{data.frame containing pedigree information as exported from tick(at)lab}

\item{features}{column names from data frame or either of: "Task", short for "HasTask","HadTask"; "Tasks", short for "HasTask","HadTask","ParticipatesInActiveMating"; NULL. All columns need to contain data convertible into logical}
}
\value{
An object of class \code{pedigree} or \code{pedigreeList}
  Containing the following items:
  famid id findex mindex sex  affected status relation
}
\description{
A convenience function for easily converting a TickAtLab table into a \link[kinship2::pedigree]{kinship2::pedigree}.
}
