% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickAtLab_SplitGenotype.R
\name{TickAtLab_SplitGenotype}
\alias{TickAtLab_SplitGenotype}
\title{TickAtLab_SplitGenotype}
\usage{
TickAtLab_SplitGenotype(df)
}
\arguments{
\item{df}{Original data.frame in the "Tick@Lab format"}
}
\value{
A data frame containing animal colony information. One row per animal. Per genotype information stored, contains one column (named by the gene) soring the genotype information.
}
\description{
Subsets a "Tick@Lab format" data.frame by line name.
}
\details{
FIXME: This function currently works with only one relevant gene.
}
