% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickAtLab_SubsetByFamily.R
\name{TickAtLab_SubsetByFamily}
\alias{TickAtLab_SubsetByFamily}
\title{TickAtLab_SubsetByFamily}
\usage{
TickAtLab_SubsetByFamily(df, Stamm)
}
\arguments{
\item{df}{Original data.frame in the "Tick@Lab format"}

\item{Stamm}{Name of the line to keep.}
}
\value{
A data frame containing animal colony information. One row per animal.
}
\description{
Subsets a "Tick@Lab format" data.frame by line name.
}
