% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickAtLab_Update.R
\name{TickAtLab_Update}
\alias{TickAtLab_Update}
\title{TickAtLab_Update}
\usage{
TickAtLab_Update(df = mousedb, df.new = TickAtLab_Import())
}
\arguments{
\item{df}{Original data.frame in the "Tick@Lab format"}

\item{df.new}{Aata.frame in the "Tick@Lab format" containing the new data}
}
\value{
A data frame containing animal colony information. One row per animal.
}
\description{
Updates a data.frame in the "Tick@Lab format" with information on new animals. Adds/Updates the "HadTask" column, containing the status of the HasTask column before the updata.
}
